task 6: fix

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -3,6 +3,6 @@ from typing import Union, List
 """ type-annotated function: sum_mixed_list """
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """ return sum of list """
-    return sum(mxd_lst)
+    return float(sum(mxd_lst))
